chore: wavelet utils and op built upon model py file loading
