+++ models/modules/utils.py
@@ -12,8 +12,6 @@ from torchvision.models import vgg
 
 from util.util import tensor2im
 
-from .op import upfirdn2d
-
 ##########################################################
 # Fonctions used for networks initialisation
 ##########################################################
@@ -301,60 +299,3 @@ def predict_depth(img, midas, model_type):
     )
     prediction = midas(transform(img))  # .unsqueeze(dim=1)
     return prediction
-
-
-### functions for frequency space
-
-
-def get_haar_wavelet(in_channels):
-    haar_wav_l = 1 / (2**0.5) * torch.ones(1, 2)
-    haar_wav_h = 1 / (2**0.5) * torch.ones(1, 2)
-    haar_wav_h[0, 0] = -1 * haar_wav_h[0, 0]
-
-    haar_wav_ll = haar_wav_l.T * haar_wav_l
-    haar_wav_lh = haar_wav_h.T * haar_wav_l
-    haar_wav_hl = haar_wav_l.T * haar_wav_h
-    haar_wav_hh = haar_wav_h.T * haar_wav_h
-
-    return haar_wav_ll, haar_wav_lh, haar_wav_hl, haar_wav_hh
-
-
-class HaarTransform(nn.Module):
-    def __init__(self, in_channels):
-        super().__init__()
-
-        ll, lh, hl, hh = get_haar_wavelet(in_channels)
-
-        self.register_buffer("ll", ll)
-        self.register_buffer("lh", lh)
-        self.register_buffer("hl", hl)
-        self.register_buffer("hh", hh)
-
-    def forward(self, input):
-        ll = upfirdn2d(input, self.ll, down=2)
-        lh = upfirdn2d(input, self.lh, down=2)
-        hl = upfirdn2d(input, self.hl, down=2)
-        hh = upfirdn2d(input, self.hh, down=2)
-
-        return torch.cat((ll, lh, hl, hh), 1)
-
-
-class InverseHaarTransform(nn.Module):
-    def __init__(self, in_channels):
-        super().__init__()
-
-        ll, lh, hl, hh = get_haar_wavelet(in_channels)
-
-        self.register_buffer("ll", ll)
-        self.register_buffer("lh", -lh)
-        self.register_buffer("hl", -hl)
-        self.register_buffer("hh", hh)
-
-    def forward(self, input):
-        ll, lh, hl, hh = input.chunk(4, 1)
-        ll = upfirdn2d(ll, self.ll, up=2, pad=(1, 0, 1, 0))
-        lh = upfirdn2d(lh, self.lh, up=2, pad=(1, 0, 1, 0))
-        hl = upfirdn2d(hl, self.hl, up=2, pad=(1, 0, 1, 0))
-        hh = upfirdn2d(hh, self.hh, up=2, pad=(1, 0, 1, 0))
-
-        return ll + lh + hl + hh
